Add discordian.sh control script with logging
This can be used to startup or restart the DiscordianAI bot.

It will pkill any existing instances of the bot before relaunching.

This is safe to use withi a crontab as well.

All logs include timestamps.

Features Include:
 -b or --background will output to nohup.out
 -d or --daemon will redirect all output to /dev/null
 -f or --folder defines the working folder (useful for crontabs)
 -c or --config to define a specific .ini file for loading

When ran without -b or -d the output will be in your terminal.

NOTE: You'll want to uncomment the PATH= if using in a crontab
      without expressly defining the working folder.