Improve Dockerfile compatibility and performance
This Dockerfile now uses the `python:3.12-slim-bullseye` image, which
has better compatibility than the alpine image. It also sets the working
directory to `/app` for better organization. The `requirements.txt` file
is copied separately and the dependencies are installed before the rest
of the code is copied over. This allows Docker to cache the installed
dependencies for faster builds when your code changes but your
dependencies do not. The `--no-cache-dir` option is used with
`pip install` to keep the image size small.

Finally, the ENTRYPOINT and CMD instructions are used to run your bot
with the specified configuration file. Note that this Dockerfile expects
the `config.ini` file to be in the same directory as the Dockerfile. If
it’s in a different location, you’ll need to adjust the CMD instruction
accordingly. Also, no ports are exposed as per your request.

When you run the Docker image, you can use the `-v` option to mount the
`config.ini` file from your host machine to the Docker container.